refactor: GetLatestTip -> GetLatestFullTip
